Feature/#727 refactor matchmaker for 4v4 party algorithm (#749)
* Split matchmaker.algorithm into stable_marriage and bucket_teams

* Make BucketTeamMatchmaker implement Matchmaker

* Make Matchmaker reusable

Move _register_unmatched_searches into matchmaker_queue, add searches as an argument to Matchmaker.find

* Clean up formatting and imports

* Clean up unit test imports, run isort again

* Add some tests for BucketTeamMatchmaker

* Let Matchmaker.find return list of unmatched searches

* Add test for failed matching attempts registered when search is canceled

and minor changes to satisfy static checkers

* Remove unused imports, unused variables

* Fix formatting

* Issue/#801: Register failed matching attempt when queue is too small

* Refactor abstract class with ABC

* Use single extend instead of two appends

* Use hypothesis.assume to filter out uninteresting test data
