+++ server/matchmaker/algorithm/random_newbies.py
@@ -1,0 +1,38 @@
+from typing import Dict, Iterable, List, Tuple
+
+from ..search import Match, Search
+from .matchmaker import MatchmakingPolicy1v1
+
+
+class RandomlyMatchNewbies(MatchmakingPolicy1v1):
+    def find(
+        self, searches: Iterable[Search]
+    ) -> Tuple[Dict[Search, Search], List[Match]]:
+        self.matches.clear()
+        searches_remaining_unmatched = set(searches)
+
+        unmatched_newbies: List[Search] = []
+        first_opponent = None
+        for search in searches:
+            if search.has_top_player():
+                continue
+            elif search.has_newbie():
+                unmatched_newbies.append(search)
+            elif not first_opponent and search.failed_matching_attempts >= 1:
+                first_opponent = search
+
+        while len(unmatched_newbies) >= 2:
+            newbie1 = unmatched_newbies.pop()
+            newbie2 = unmatched_newbies.pop()
+            self._match(newbie1, newbie2)
+            searches_remaining_unmatched.discard(newbie1)
+            searches_remaining_unmatched.discard(newbie2)
+
+        can_match_last_newbie_with_first_opponent = unmatched_newbies and first_opponent
+        if can_match_last_newbie_with_first_opponent:
+            newbie = unmatched_newbies[0]
+            self._match(newbie, first_opponent)
+            searches_remaining_unmatched.discard(newbie)
+            searches_remaining_unmatched.discard(first_opponent)
+
+        return self.matches, list(searches_remaining_unmatched)
