+++ tests/unit_tests/test_matchmaker_algorithm_stable_marriage.py
@@ -1,14 +1,11 @@
 import logging
-import math
-import random
 
 import pytest
-from hypothesis import given, settings
-from hypothesis import strategies as st
+from hypothesis import assume, given, settings
 
 from server import config
-from server.matchmaker import Search, algorithm
-from server.rating import RatingType
+from server.matchmaker import Search
+from server.matchmaker.algorithm import stable_marriage
 
 from .strategies import st_searches_list
 
@@ -32,7 +29,7 @@ def player_factory(player_factory):
     return make
 
 
-def add_graph_edge_weights(graph) -> algorithm.WeightedGraph:
+def add_graph_edge_weights(graph) -> stable_marriage.WeightedGraph:
     return {
         s1: [(s2, s1.quality_with(s2)) for s2 in edges]
         for s1, edges in graph.items()
@@ -40,8 +37,8 @@ def add_graph_edge_weights(graph) -> algorithm.WeightedGraph:
 
 
 @pytest.mark.parametrize("build_func", (
-    algorithm._MatchingGraph.build_full,
-    algorithm._MatchingGraph.build_fast
+    stable_marriage._MatchingGraph.build_full,
+    stable_marriage._MatchingGraph.build_fast
 ))
 def test_build_full_matching_graph(player_factory, build_func):
     # For small numbers of searches, build_full and build_fast should create
@@ -61,8 +58,8 @@ def test_build_full_matching_graph(player_factory, build_func):
 
 
 @pytest.mark.parametrize("build_func", (
-    algorithm._MatchingGraph.build_full,
-    algorithm._MatchingGraph.build_fast
+    stable_marriage._MatchingGraph.build_full,
+    stable_marriage._MatchingGraph.build_fast
 ))
 def test_build_matching_graph_different_ranks(player_factory, build_func):
     s1 = Search([player_factory(1500, 64, ladder_games=20)])
@@ -89,7 +86,7 @@ def test_remove_isolated(player_factory):
         s3: [s1]
     })
 
-    algorithm._MatchingGraph.remove_isolated(ranks)
+    stable_marriage._MatchingGraph.remove_isolated(ranks)
 
     assert ranks == add_graph_edge_weights({
         s1: [s3],
@@ -107,14 +104,14 @@ def test_remove_isolated_2(player_factory):
         s3: []
     }
 
-    algorithm._MatchingGraph.remove_isolated(ranks)
+    stable_marriage._MatchingGraph.remove_isolated(ranks)
 
     assert ranks == {}
 
 
 @pytest.mark.parametrize("build_func", (
-    algorithm._MatchingGraph.build_full,
-    algorithm._MatchingGraph.build_fast
+    stable_marriage._MatchingGraph.build_full,
+    stable_marriage._MatchingGraph.build_fast
 ))
 def test_match_graph_will_not_include_matches_below_threshold_quality(player_factory, build_func):
     s1 = Search([player_factory(1500, 500)])
@@ -130,8 +127,8 @@ def test_match_graph_will_not_include_matches_below_threshold_quality(player_fac
 
 
 @pytest.mark.parametrize("build_func", (
-    algorithm._MatchingGraph.build_full,
-    algorithm._MatchingGraph.build_fast
+    stable_marriage._MatchingGraph.build_full,
+    stable_marriage._MatchingGraph.build_fast
 ))
 @given(searches=st_searches_list(max_players=2))
 @settings(deadline=300)
@@ -153,8 +150,8 @@ def test_matching_graph_symmetric(
 
 
 @pytest.mark.parametrize("build_func", (
-    algorithm._MatchingGraph.build_full,
-    algorithm._MatchingGraph.build_fast
+    stable_marriage._MatchingGraph.build_full,
+    stable_marriage._MatchingGraph.build_fast
 ))
 @given(searches=st_searches_list(max_players=2))
 @settings(deadline=300)
@@ -169,7 +166,9 @@ def test_stable_marriage_produces_symmetric_matchings(
 
         ranks = build_func(searches)
 
-        matches = algorithm.StableMarriage().find(ranks)
+        matches = stable_marriage.StableMarriage().find(ranks)
+
+        assume(matches != [])
 
         for search in matches:
             opponent = matches[search]
@@ -185,9 +184,9 @@ def test_stable_marriage(player_factory):
     s6 = Search([player_factory(1250, 175, name="p6")])
 
     searches = [s1, s2, s3, s4, s5, s6]
-    ranks = algorithm._MatchingGraph.build_full(searches)
+    ranks = stable_marriage._MatchingGraph.build_full(searches)
 
-    matches = algorithm.StableMarriage().find(ranks)
+    matches = stable_marriage.StableMarriage().find(ranks)
 
     assert matches[s1] == s4
     assert matches[s2] == s5
@@ -201,9 +200,9 @@ def test_stable_marriage_matches_new_players_with_new_and_old_with_old_if_differ
     old2 = Search([player_factory(2350, 75, name="old2", ladder_games=200)])
 
     searches = [new1, new2, old1, old2]
-    ranks = algorithm._MatchingGraph.build_full(searches)
+    ranks = stable_marriage._MatchingGraph.build_full(searches)
 
-    matches = algorithm.StableMarriage().find(ranks)
+    matches = stable_marriage.StableMarriage().find(ranks)
 
     assert matches[new1] == new2
     assert matches[old1] == old2
@@ -218,9 +217,9 @@ def test_stable_marriage_matches_new_players_with_new_and_old_with_old_if_same_m
     old2 = Search([player_factory(500, 75, name="old2", ladder_games=100)])
 
     searches = [new1, new2, old1, old2]
-    ranks = algorithm._MatchingGraph.build_full(searches)
+    ranks = stable_marriage._MatchingGraph.build_full(searches)
 
-    matches = algorithm.StableMarriage().find(ranks)
+    matches = stable_marriage.StableMarriage().find(ranks)
 
     assert matches[new1] == new2
     assert matches[old1] == old2
@@ -235,9 +234,9 @@ def test_stable_marriage_better_than_greedy(player_factory):
     s6 = Search([player_factory(2400, 64, name="p6")])
 
     searches = [s1, s2, s3, s4, s5, s6]
-    ranks = algorithm._MatchingGraph.build_full(searches)
+    ranks = stable_marriage._MatchingGraph.build_full(searches)
 
-    matches = algorithm.StableMarriage().find(ranks)
+    matches = stable_marriage.StableMarriage().find(ranks)
 
     # Note that the most balanced configuration would be
     # (s1, s6)  quality: 0.93
@@ -258,9 +257,9 @@ def test_stable_marriage_unmatch(player_factory):
     s4 = Search([player_factory(505, 64, name="p4")])
 
     searches = [s1, s2, s3, s4]
-    ranks = algorithm._MatchingGraph.build_full(searches)
+    ranks = stable_marriage._MatchingGraph.build_full(searches)
 
-    matches = algorithm.StableMarriage().find(ranks)
+    matches = stable_marriage.StableMarriage().find(ranks)
 
     assert matches[s1] == s4  # quality: 0.96622
     assert matches[s2] == s3  # quality: 0.96623
@@ -275,9 +274,10 @@ def test_random_newbie_matching_is_symmetric(player_factory):
     s6 = Search([player_factory(600, 500, name="p6", ladder_games=5)])
 
     searches = [s1, s2, s3, s4, s5, s6]
-    matches = algorithm.RandomlyMatchNewbies().find(searches)
+    matches, unmatched_searches = stable_marriage.RandomlyMatchNewbies().find(searches)
 
-    assert matches
+    assert len(matches) == len(searches)
+    assert len(unmatched_searches) == 0
 
     for search in matches:
         opponent = matches[search]
@@ -291,10 +291,11 @@ def test_newbies_are_forcefully_matched_with_newbies(player_factory):
     pro.register_failed_matching_attempt()
 
     searches = [newbie1, pro, newbie2]
-    matches = algorithm.RandomlyMatchNewbies().find(searches)
+    matches, unmatched_searches = stable_marriage.RandomlyMatchNewbies().find(searches)
 
     assert matches[newbie1] == newbie2
     assert matches[newbie2] == newbie1
+    assert unmatched_searches == [pro]
 
 
 def test_newbie_team_matched_with_newbie_team(player_factory):
@@ -308,10 +309,11 @@ def test_newbie_team_matched_with_newbie_team(player_factory):
     ])
 
     searches = [newbie1, newbie2]
-    matches = algorithm.RandomlyMatchNewbies().find(searches)
+    matches, unmatched_searches = stable_marriage.RandomlyMatchNewbies().find(searches)
 
     assert matches[newbie1] == newbie2
     assert matches[newbie2] == newbie1
+    assert len(unmatched_searches) == 0
 
 
 def test_partial_newbie_team_matched_with_newbie_team(player_factory):
@@ -325,10 +327,11 @@ def test_partial_newbie_team_matched_with_newbie_team(player_factory):
     ])
 
     searches = [partial_newbie, newbie]
-    matches = algorithm.RandomlyMatchNewbies().find(searches)
+    matches, unmatched_searches = stable_marriage.RandomlyMatchNewbies().find(searches)
 
     assert matches[partial_newbie] == newbie
     assert matches[newbie] == partial_newbie
+    assert len(unmatched_searches) == 0
 
 
 def test_newbie_and_top_rated_team_not_matched_randomly(player_factory):
@@ -342,9 +345,10 @@ def test_newbie_and_top_rated_team_not_matched_randomly(player_factory):
     ])
 
     searches = [newbie_and_top_rated, newbie]
-    matches = algorithm.RandomlyMatchNewbies().find(searches)
+    matches, unmatched_searches = stable_marriage.RandomlyMatchNewbies().find(searches)
 
     assert not matches
+    assert len(unmatched_searches) == len(searches)
 
 
 def test_unmatched_newbies_forcefully_match_pros(player_factory):
@@ -352,13 +356,15 @@ def test_unmatched_newbies_forcefully_match_pros(player_factory):
     pro = Search([player_factory(1400, 10, ladder_games=100)])
 
     searches = [newbie, pro]
-    matches = algorithm.RandomlyMatchNewbies().find(searches)
+    matches, unmatched_searches = stable_marriage.RandomlyMatchNewbies().find(searches)
     # No match if the pro is on their first attempt
     assert len(matches) == 0
+    assert len(unmatched_searches) == 2
 
     pro.register_failed_matching_attempt()
-    matches = algorithm.RandomlyMatchNewbies().find(searches)
+    matches, unmatched_searches = stable_marriage.RandomlyMatchNewbies().find(searches)
     assert len(matches) == 2
+    assert len(unmatched_searches) == 0
 
 
 def test_newbie_team_matched_with_pro_team(player_factory):
@@ -372,13 +378,15 @@ def test_newbie_team_matched_with_pro_team(player_factory):
     ])
 
     searches = [newbie, pro]
-    matches = algorithm.RandomlyMatchNewbies().find(searches)
+    matches, unmatched_searches = stable_marriage.RandomlyMatchNewbies().find(searches)
     # No match if the pros are on their first attempt
     assert len(matches) == 0
+    assert len(unmatched_searches) == 2
 
     pro.register_failed_matching_attempt()
-    matches = algorithm.RandomlyMatchNewbies().find(searches)
+    matches, unmatched_searches = stable_marriage.RandomlyMatchNewbies().find(searches)
     assert len(matches) == 2
+    assert len(unmatched_searches) == 0
 
 
 def test_unmatched_newbies_do_not_forcefully_match_top_players(player_factory):
@@ -387,9 +395,10 @@ def test_unmatched_newbies_do_not_forcefully_match_top_players(player_factory):
     top_player.register_failed_matching_attempt()
 
     searches = [newbie, top_player]
-    matches = algorithm.RandomlyMatchNewbies().find(searches)
+    matches, unmatched_searches = stable_marriage.RandomlyMatchNewbies().find(searches)
 
     assert len(matches) == 0
+    assert len(unmatched_searches) == 2
 
 
 def test_newbie_team_dos_not_match_with_top_players_team(player_factory):
@@ -404,9 +413,10 @@ def test_newbie_team_dos_not_match_with_top_players_team(player_factory):
     top_player.register_failed_matching_attempt()
 
     searches = [newbie, top_player]
-    matches = algorithm.RandomlyMatchNewbies().find(searches)
+    matches, unmatched_searches = stable_marriage.RandomlyMatchNewbies().find(searches)
 
     assert len(matches) == 0
+    assert len(unmatched_searches) == 2
 
 
 def unmatched_newbie_teams_do_not_forcefully_match_pros(player_factory):
@@ -418,9 +428,10 @@ def unmatched_newbie_teams_do_not_forcefully_match_pros(player_factory):
     pro.register_failed_matching_attempt()
 
     searches = [newbie_team, pro]
-    matches = algorithm.RandomlyMatchNewbies().find(searches)
+    matches, unmatched_searches = stable_marriage.RandomlyMatchNewbies().find(searches)
 
     assert len(matches) == 0
+    assert len(unmatched_searches) == 2
 
 
 def test_odd_number_of_unmatched_newbies(player_factory):
@@ -431,9 +442,10 @@ def test_odd_number_of_unmatched_newbies(player_factory):
     pro.register_failed_matching_attempt()
 
     searches = [newbie1, pro, newbie2, newbie3]
-    matches = algorithm.RandomlyMatchNewbies().find(searches)
+    matches, unmatched_searches = stable_marriage.RandomlyMatchNewbies().find(searches)
 
     assert len(matches) == 4
+    assert len(unmatched_searches) == 0
 
 
 def test_matchmaker(player_factory):
@@ -460,12 +472,15 @@ def test_matchmaker(player_factory):
         pro_alone,
         top_player
     ]
-    match_pairs = algorithm.make_matches(searches)
+    team_size = 1
+    matchmaker = stable_marriage.StableMarriageMatchmaker()
+    match_pairs, unmatched_searches = matchmaker.find(searches, team_size)
     match_sets = [set(pair) for pair in match_pairs]
 
     assert {newbie_that_matches1, newbie_that_matches2} in match_sets
     assert {pro_that_matches1, pro_that_matches2} in match_sets
     assert {newbie_force_matched, pro_alone} in match_sets
+    assert unmatched_searches == [top_player]
     for match_pair in match_pairs:
         assert top_player not in match_pair
 
@@ -478,7 +493,9 @@ def test_matchmaker_performance(player_factory, bench, caplog):
     searches = [Search([player_factory(1500, 500, ladder_games=1)]) for _ in range(NUM_SEARCHES)]
 
     with bench:
-        algorithm.make_matches(searches)
+        team_size = 1
+        matchmaker = stable_marriage.StableMarriageMatchmaker()
+        matchmaker.find(searches, team_size)
 
     assert bench.elapsed() < 0.5
 
@@ -488,248 +505,76 @@ def test_matchmaker_random_only(player_factory):
     newbie2 = Search([player_factory(200, 400, ladder_games=9)])
 
     searches = (newbie1, newbie2)
-    match_pairs = algorithm.make_matches(searches)
+    team_size = 1
+    matchmaker = stable_marriage.StableMarriageMatchmaker()
+    match_pairs, unmatched_searches = matchmaker.find(searches, team_size)
     match_sets = [set(pair) for pair in match_pairs]
 
     assert {newbie1, newbie2} in match_sets
+    assert len(unmatched_searches) == 0
 
 
-def test_make_matches_will_not_match_low_quality_games(player_factory):
+def test_find_will_not_match_low_quality_games(player_factory):
     s1 = Search([player_factory(100, 64, name="p1")])
     s2 = Search([player_factory(2000, 64, name="p2")])
 
     searches = [s1, s2]
 
-    matches = algorithm.make_matches(searches)
+    team_size = 1
+    matchmaker = stable_marriage.StableMarriageMatchmaker()
+    matches, unmatched_searches = matchmaker.find(searches, team_size)
 
-    assert (s1, s2) not in matches
-    assert (s2, s1) not in matches
-
-
-def test_make_matches_communicates_failed_attempts(player_factory):
-    s1 = Search([player_factory(100, 64, name="p1")])
-    s2 = Search([player_factory(2000, 64, name="p2")])
-
-    searches = [s1, s2]
-
-    assert s1.failed_matching_attempts == 0
-    assert s2.failed_matching_attempts == 0
-
-    matches = algorithm.make_matches(searches)
-
-    # These searches should not have been matched
-    assert not matches
-    assert s1.failed_matching_attempts == 1
-    assert s2.failed_matching_attempts == 1
+    assert len(matches) == 0
+    assert len(unmatched_searches) == len(searches)
 
 
-@pytest.mark.parametrize("make_teams_func", (
-    algorithm.make_teams,
-    algorithm.make_teams_from_single
-))
-@given(
-    searches=st_searches_list(max_players=1),
-    size=st.integers(min_value=1, max_value=10)
-)
-def test_make_teams_single_correct_size(searches, size, make_teams_func):
-    matched, non_matched = make_teams_func(searches, size)
-
-    for search in matched:
-        assert len(search.players) == size
-
-
-def test_make_teams_single_2v2_large_pool(player_factory):
-    """
-    When we have a large number of players all with similar ratings, we want
-    teams to be formed by putting players with the same rating on the same team.
-    """
-
-    # Large enough so the test is unlikely to pass by chance
-    num = 40
-
-    searches = [
-        Search([player_factory(random.uniform(950, 1050), 10, name=f"p{i}")])
-        for i in range(num)
+def test_unmatched_searches_without_newbies(player_factory):
+    players = [
+        player_factory(100, 10, name="lowRating_unmatched_1"),
+        player_factory(500, 10, name="lowRating_unmatched_2"),
+        player_factory(1500, 10, name="midRating_matched_1"),
+        player_factory(1500, 10, name="midRating_matched_2"),
+        player_factory(1500, 10, name="midRating_matched_3"),
+        player_factory(1500, 10, name="midRating_matched_4"),
+        player_factory(2000, 10, name="highRating_unmatched_1"),
+        player_factory(2500, 10, name="highRating_unmatched_2"),
     ]
-    searches += [
-        Search([player_factory(random.uniform(450, 550), 10, name=f"p{i}")])
-        for i in range(num)
+    searches = [Search([player]) for player in players]
+
+    team_size = 1
+    matchmaker = stable_marriage.StableMarriageMatchmaker()
+    matches, unmatched_searches = matchmaker.find(searches, team_size)
+
+    expected_number_of_matches = 2
+    assert len(matches) == expected_number_of_matches
+    assert len(unmatched_searches) == len(searches) - 2 * team_size * expected_number_of_matches
+
+
+def test_unmatched_searches_with_newbies(player_factory):
+    players = [
+        player_factory(100, 10, name="newbie1", ladder_games=1),
+        player_factory(200, 10, name="newbie2", ladder_games=1),
+        player_factory(300, 10, name="newbie3", ladder_games=1),
+        player_factory(400, 10, name="newbie4", ladder_games=1),
+        player_factory(500, 10, name="newbie5", ladder_games=1),
+        player_factory(1500, 10, name="midRating_matched_1"),
+        player_factory(1500, 10, name="midRating_matched_2"),
+        player_factory(1500, 10, name="midRating_matched_3"),
+        player_factory(1500, 10, name="midRating_matched_4"),
+        player_factory(2000, 10, name="highRating_unmatched_1"),
+        player_factory(2500, 10, name="highRating_unmatched_2"),
     ]
-    matched, non_matched = algorithm.make_teams_from_single(searches, size=2)
+    searches = [Search([player]) for player in players]
 
-    assert matched != []
-    assert non_matched == []
+    force_matched_player = player_factory(750, 10, name="lowRating_unmatched_1")
+    force_matched_search = Search([force_matched_player])
+    force_matched_search.register_failed_matching_attempt()
+    searches.append(force_matched_search)
 
-    for search in matched:
-        p1, p2 = search.players
-        p1_mean, _ = p1.ratings[RatingType.LADDER_1V1]
-        p2_mean, _ = p2.ratings[RatingType.LADDER_1V1]
-        #
-        assert math.fabs(p1_mean - p2_mean) <= 100
+    team_size = 1
+    matchmaker = stable_marriage.StableMarriageMatchmaker()
+    matches, unmatched_searches = matchmaker.find(searches, team_size)
 
-
-def test_make_teams_single_2v2_small_pool(player_factory):
-    """
-    When we have a small number of players, we want teams to be formed by
-    distributing players of equal skill to different teams so that we can
-    maximize the chances of getting a match.
-    """
-
-    # Try a bunch of times so it is unlikely to pass by chance
-    for _ in range(20):
-        searches = [
-            Search(
-                [player_factory(random.gauss(1000, 5), 10, name=f"p{i}")]
-            ) for i in range(2)
-        ]
-        searches += [
-            Search(
-                [player_factory(random.gauss(500, 5), 10, name=f"r{i}")]
-            ) for i in range(2)
-        ]
-        matched, non_matched = algorithm.make_teams_from_single(searches, size=2)
-
-        assert matched != []
-        assert non_matched == []
-
-        for search in matched:
-            p1, p2 = search.players
-            # Order doesn't matter
-            if p1.ratings[RatingType.LADDER_1V1][0] > 900:
-                assert p2.ratings[RatingType.LADDER_1V1][0] < 600
-            else:
-                assert p1.ratings[RatingType.LADDER_1V1][0] < 600
-                assert p2.ratings[RatingType.LADDER_1V1][0] > 900
-
-
-def test_make_buckets_performance(bench, player_factory):
-    NUM_SEARCHES = 1000
-    searches = [
-        Search([player_factory(
-            random.gauss(1500, 200),
-            500,
-            ladder_games=1
-        )])
-        for _ in range(NUM_SEARCHES)
-    ]
-
-    with bench:
-        algorithm._make_buckets(searches)
-
-    assert bench.elapsed() < 0.15
-
-
-def test_make_teams_1(player_factory):
-    teams = [
-        [player_factory(name="p1"), player_factory(name="p2"), player_factory(name="p3")],
-        [player_factory(name="p4"), player_factory(name="p5")],
-        [player_factory(name="p6"), player_factory(name="p7")],
-        [player_factory(name="p8")],
-        [player_factory(name="p9")],
-        [player_factory(name="p10")],
-        [player_factory(name="p11")]
-    ]
-    do_test_make_teams(
-        teams,
-        team_size=3,
-        total_unmatched=2,
-        unmatched_sizes={1}
-    )
-
-
-def test_make_teams_2(player_factory):
-    teams = [
-        [player_factory(name="p1"), player_factory(name="p2"), player_factory(name="p3")],
-        [player_factory(name="p4"), player_factory(name="p5")],
-        [player_factory(name="p6"), player_factory(name="p7")],
-        [player_factory(name="p8")],
-        [player_factory(name="p9")],
-        [player_factory(name="p10")],
-        [player_factory(name="p11")]
-    ]
-    do_test_make_teams(
-        teams,
-        team_size=2,
-        total_unmatched=1,
-        unmatched_sizes={3}
-    )
-
-
-def test_make_teams_3(player_factory):
-    teams = [
-        [player_factory(name=f"p{i+1}")] for i in range(9)
-    ]
-    do_test_make_teams(
-        teams,
-        team_size=4,
-        total_unmatched=1,
-        unmatched_sizes={1}
-    )
-
-
-def test_make_teams_4(player_factory):
-    teams = [[player_factory()] for i in range(9)]
-    teams += [[player_factory(), player_factory()] for i in range(5)]
-    teams += [[player_factory(), player_factory(), player_factory()] for i in range(15)]
-    teams += [[player_factory(), player_factory(), player_factory(), player_factory()] for i in range(4)]
-    do_test_make_teams(
-        teams,
-        team_size=4,
-        total_unmatched=7,
-        unmatched_sizes={3, 2}
-    )
-
-
-def test_make_teams_5(player_factory):
-    teams = [
-        [player_factory(name="p1"), player_factory(name="p2"), player_factory(name="p3")],
-        [player_factory(name="p4"), player_factory(name="p5")],
-        [player_factory(name="p6"), player_factory(name="p7")],
-    ]
-    do_test_make_teams(
-        teams,
-        team_size=4,
-        total_unmatched=1,
-        unmatched_sizes={3}
-    )
-
-
-def do_test_make_teams(teams, team_size, total_unmatched, unmatched_sizes):
-    searches = [Search(t) for t in teams]
-
-    matched, non_matched = algorithm.make_teams(searches, size=team_size)
-    players_non_matched = [s.players for s in non_matched]
-
-    for s in matched:
-        assert len(s.players) == team_size
-    assert len(players_non_matched) == total_unmatched
-    for players in players_non_matched:
-        assert len(players) in unmatched_sizes
-
-
-def test_distribute_pairs_1(player_factory):
-    players = [player_factory(1500, 500, name=f"p{i+1}") for i in range(4)]
-    searches = [(Search([player]), 0) for player in players]
-    p1, p2, p3, p4 = players
-
-    grouped = [search.players for search in algorithm._distribute(searches, 2)]
-    assert grouped == [[p1, p4], [p2, p3]]
-
-
-def test_distribute_pairs_2(player_factory):
-    players = [player_factory(1500, 500, name=f"p{i+1}") for i in range(8)]
-    searches = [(Search([player]), 0) for player in players]
-    p1, p2, p3, p4, p5, p6, p7, p8 = players
-
-    grouped = [search.players for search in algorithm._distribute(searches, 2)]
-    assert grouped == [[p1, p4], [p2, p3], [p5, p8], [p6, p7]]
-
-
-def test_distribute_triples(player_factory):
-    players = [player_factory(1500, 500, name=f"p{i+1}") for i in range(6)]
-    searches = [(Search([player]), 0) for player in players]
-    p1, p2, p3, p4, p5, p6 = players
-
-    grouped = [search.players for search in algorithm._distribute(searches, 3)]
-
-    assert grouped == [[p1, p3, p6], [p2, p4, p5]]
+    expected_number_of_matches = 5
+    assert len(matches) == expected_number_of_matches
+    assert len(unmatched_searches) == len(searches) - 2 * team_size * expected_number_of_matches
